refactor(bpa-server): extract the server lifecycle into a BpaServer type
